requirements.txt
altgraph==0.17.5
blinker==1.9.0
bottle==0.13.4
cffi==2.0.0
click==8.3.1
clr_loader==0.2.10
colorama==0.4.6
Flask==3.1.3
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pefile==2024.8.26
proxy_tools==0.1.0
pycparser==3.0
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.3
pythonnet==3.0.5
pywebview==6.1
pywin32-ctypes==0.2.3
typing_extensions==4.15.0
Werkzeug==3.1.7
yt-dlp==2026.3.17
